custom-gpt-python/requirements.txt
openai
llama-index
google-auth-oauthlib
redis
python-dotenv